E_shop/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-shopping-cart==0.1
idna==3.4
paypal==1.2.5
Pillow==9.5.0
requests==2.30.0
sqlparse==0.4.4
stripe==5.4.0
tzdata==2023.3
urllib3==2.0.2
